requirements.txt
flask
newsapi-python
flask_googlecharts
alpha_vantage
google.cloud
werkzeug
nltk